refactor: integrate URL globbing, add ADR doc references, and split large modules
