+++ aura-core/src/piece_picker/logic.rs
@@ -7,6 +7,7 @@ use std::collections::HashMap;
 pub type PeerId = [u8; 20];
 
 /// Manages piece selection logic, prioritizing the rarest pieces in the swarm.
+/// Implements the BitTorrent Endgame Mode strategy detailed in [ADR-0039](aura-docs/adr/0039-bittorrent-endgame-mode.md).
 #[derive(Debug)]
 pub struct PiecePicker {
     pub num_pieces: usize,
