+++ aura-core/src/task/mod.rs
@@ -4,9 +4,11 @@ pub mod extension;
 pub mod meta;
 pub mod phase;
 pub mod range;
+pub mod state;
 pub mod subtask;
 
-pub use meta::{MetaTask, TaskState};
+pub use meta::MetaTask;
 pub use phase::{DownloadPhase, FollowOnAction, TaskType};
 pub use range::Range;
+pub use state::TaskState;
 pub use subtask::SubTask;
